requirements.txt
fastapi
sqlmodel
uvicorn
FastAPI-SQLAlchemy
psycopg2-binary
python-multipart
python-jose[cryptography]
passlib[bcrypt]